+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to
 
 ### Fixed
 
+- Fixed an `EmptyCellVertex` data integrity issue between `AddressMapping` and `DependencyGraph` objects.
+  [#1188](https://github.com/handsontable/hyperformula/issues/1188)
 - Fixed build issue for developers using M1 and M2-chip MacBooks. 
   [#1166](https://github.com/handsontable/hyperformula/issues/1166)
 - Fixed an issue where order of items in the result of `removeColumns()` method depended on the address mapping policy.